Add sample-wise gradients clipping to `TorchModel`.
Note: to efficiently compute (and clip) sample-wise gradients
in Torch, the additional dependency `functorch` is used. This
poses a risk as to the maintainance of `declearn[torch]` due
to `functorch` still being presented as experimental by Meta;
however it is the best available solution to limit overheads,
and will probably become an important component of the torch
ecosystem in the future.