Turn `Vector` operator class attributes into properties.
The rationale behind this change is that assigning a function
to a class attribute results in it being bound as a method as
per the Python data model.
However, for unclear reasons that may have to do with the way
they are built, `operator.[op]`, `numpy.[op]` and `torch.[op]`
are not affected by this binding, which is not desirable in
the context of `Vector`.
Yet, to prevent this issue in future framework-specific classes
(which was hackily dealt with for tensorflow), and to improve
their documentation, these base operator attributes were turned
into properties.
Additionally, this commit extends the docstring of `Vector` to
provide developers with indications as to what should or may be
implemented, extended or overridden by concrete subclasses.