Reformat extra dependencies' specifications.
The previous version specified some recursive dependencies, to avoid
writing redundant dependency specifications. This is convenient, and
supported (as of versions 19.0 to 22.3) by pip. However, this is no
longer working with tox v4.0.2 (as opposed to v3.27), that was just
released.

As a consequence, and since self-dependency is not part of PEP 631,
this commit removes it, and introduces some (documented) redundancy
in extra dependencies - which as of now make for a couple of lines
only.