Drop support for Python 3.7.
Support for Python 3.7 had been envisioned and partially documented,
but had not been tested for a while. As it turns out, some of the
`typing` functionalities the generic TOML-parsing system relies upon
were only introduced in Python 3.8. Since rewriting this code would
be tedious without these utils, and since official support for py37
ends in less than six months, it was decided not to support it for
declearn.