Implement custom 'PoissonSampler' for 'TorchDataset'.
Previously, poisson sampling was delegated to a sampler implemented
as part of the third-party 'opacus' library. However, the latter is
not mandatory for end-users, and its implementation does not prevent
empty batches from being yielded (which break our collate function).

Therefore, this commit implements our own sampler, as it is a very
straightforward process.