Implement group-fairness functions.
This implementation is adapted from the one that was originally
developed on the legacy 'fairgrad' branch. It changes the design
to switch from a single 'FairnessFunction' endpoint with multiple
constants-computation backends, relying on in a hard-coded subset
of available functions, to an OOP design where 'FairnessFunction'
is an ABC, with subclasses for each implemented definition, and
the possibility to extend supported definitions with custom child
classes.