Implement 'FairnessDataset' and 'FairnessInMemoryDataset'.
This implementation was developed iteratively on the 'fairgrad'
branch, and further reworked for this commit. As its history is
complex, it is newly-committed rather than squashed out of past
commits.

The main novelty compared to the "legacy" fairgrad version it that
sensitive attributes are now designed to be specified at init time
rather than via a post-instantiation 'set_sensitive_data' call.