Implement Fed-FairBatch / FedFB algorithms for group fairness.
This commit adds the Fed-FairBatch and FedFB algorithms for group
fairness. It is based on the "legacy" implementation from the
'fairgrad' branch, but adjusts to the updated fairness API.

It also adds some documentation, changes the signature of sampling
probabilities controllers to be coherent with 'FairnessFunction',
and integrates FedFB as a variant of Fed-FairGrad, within the same
controller class.

Co-authored-by: Brahim Erraji <brahim.erraji@inria.fr>