Implement FairFed algorithm for group fairness.
This implementation is very loosely based on the "legacy" one from
the 'fairgrad' branch. It generalizes to broader cases than those
showcased in the initial FairFed paper, and relies on a dedicated
(mostly client-side) Aggregator subclass that is altered during
fairness rounds to use fairness-based averaging weights.

The maths to the "global" fairness measures computation are not
those of the original paper, and should be verified to equate the
original paper for the EOD and SPD cases  prior to merging.